chore: remove deprecated filter implementation
