+++ packages/interfaces/src/filter.ts
@@ -1,17 +1,77 @@
-import type { PeerId } from "@libp2p/interface";
-
 import type { IDecodedMessage, IDecoder } from "./message.js";
-import type { ContentTopic, ThisOrThat } from "./misc.js";
-import type {
-  Callback,
-  ProtocolError,
-  SDKProtocolResult
-} from "./protocols.js";
-import type { IReceiver } from "./receiver.js";
+import type { Callback } from "./protocols.js";
 
-export type SubscriptionCallback<T extends IDecodedMessage> = {
-  decoders: IDecoder<T>[];
-  callback: Callback<T>;
+export type IFilter = {
+  readonly multicodec: string;
+
+  /**
+   * Subscribes to messages with specified decoders and executes callback when a message is received.
+   * In case no peers available initially - will delay subscription till connects to any peer.
+   *
+   * @param decoders - Single decoder or array of decoders to subscribe to. All decoders must share the same pubsubTopic.
+   * @param callback - Function called when a message matching the decoder's contentTopic is received.
+   * @returns Promise that resolves to true if subscription was successful, false otherwise.
+   *
+   * @example
+   * // Subscribe to a single content topic
+   * await filter.subscribe(decoder, (msg) => console.log(msg));
+   *
+   * @example
+   * // Subscribe to multiple content topics with the same pubsub topic
+   * await filter.subscribe([decoder1, decoder2], (msg) => console.log(msg));
+   *
+   * @example
+   * // Handle subscription failure
+   * const success = await filter.subscribe(decoder, handleMessage);
+   * if (!success) {
+   *   console.error("Failed to subscribe");
+   * }
+   */
+  subscribe<T extends IDecodedMessage>(
+    decoders: IDecoder<T> | IDecoder<T>[],
+    callback: Callback<T>
+  ): Promise<boolean>;
+
+  /**
+   * Unsubscribes from messages with specified decoders.
+   *
+   * @param decoders - Single decoder or array of decoders to unsubscribe from. All decoders must share the same pubsubTopic.
+   * @returns Promise that resolves to true if unsubscription was successful, false otherwise.
+   *
+   * @example
+   * // Unsubscribe from a single decoder
+   * await filter.unsubscribe(decoder);
+   *
+   * @example
+   * // Unsubscribe from multiple decoders at once
+   * await filter.unsubscribe([decoder1, decoder2]);
+   *
+   * @example
+   * // Handle unsubscription failure
+   * const success = await filter.unsubscribe(decoder);
+   * if (!success) {
+   *   console.error("Failed to unsubscribe");
+   * }
+   */
+  unsubscribe<T extends IDecodedMessage>(
+    decoders: IDecoder<T> | IDecoder<T>[]
+  ): Promise<boolean>;
+
+  /**
+   * Unsubscribes from all active subscriptions across all pubsub topics.
+   *
+   * @example
+   * // Clean up all subscriptions when React component unmounts
+   * useEffect(() => {
+   *   return () => filter.unsubscribeAll();
+   * }, [filter]);
+   *
+   * @example
+   * // Reset subscriptions and start over
+   * filter.unsubscribeAll();
+   * await filter.subscribe(newDecoder, newCallback);
+   */
+  unsubscribeAll(): void;
 };
 
 export type FilterProtocolOptions = {
@@ -30,52 +90,9 @@ export type FilterProtocolOptions = {
   pingsBeforePeerRenewed: number;
 
   /**
-   * Enables js-waku to send probe LightPush message over subscribed pubsubTopics on created subscription.
-   * In case message won't be received back through Filter - js-waku will attempt to subscribe to another peer.
+   * Number of peers to be used for establishing subscriptions.
    *
-   * @default false
+   * @default 2
    */
-  enableLightPushFilterCheck: boolean;
+  numPeersToUse: number;
 };
-
-export interface ISubscription {
-  subscribe<T extends IDecodedMessage>(
-    decoders: IDecoder<T> | IDecoder<T>[],
-    callback: Callback<T>
-  ): Promise<SDKProtocolResult>;
-
-  unsubscribe(contentTopics: ContentTopic[]): Promise<SDKProtocolResult>;
-
-  ping(peerId?: PeerId): Promise<SDKProtocolResult>;
-
-  unsubscribeAll(): Promise<SDKProtocolResult>;
-}
-
-export type IFilter = IReceiver & {
-  readonly multicodec: string;
-  subscribe<T extends IDecodedMessage>(
-    decoders: IDecoder<T> | IDecoder<T>[],
-    callback: Callback<T>
-  ): Promise<SubscribeResult>;
-};
-
-export type SubscribeResult = SubscriptionSuccess | SubscriptionError;
-
-type SubscriptionSuccess = {
-  subscription: ISubscription;
-  error: null;
-  results: SDKProtocolResult;
-};
-
-type SubscriptionError = {
-  subscription: null;
-  error: ProtocolError;
-  results: null;
-};
-
-export type CreateSubscriptionResult = ThisOrThat<
-  "subscription",
-  ISubscription,
-  "error",
-  ProtocolError
->;
