+++ packages/interfaces/src/receiver.ts
@@ -1,14 +1,10 @@
 import type { IDecodedMessage, IDecoder } from "./message.js";
-import type {
-  ContentTopic,
-  IAsyncIterator,
-  PubsubTopic,
-  Unsubscribe
-} from "./misc.js";
+import type { IAsyncIterator, Unsubscribe } from "./misc.js";
 import type { Callback } from "./protocols.js";
 
-export type ActiveSubscriptions = Map<PubsubTopic, ContentTopic[]>;
-
+/**
+ * @deprecated will be replaced in next version
+ */
 export interface IReceiver {
   toSubscriptionIterator: <T extends IDecodedMessage>(
     decoders: IDecoder<T> | IDecoder<T>[]
