+++ packages/interfaces/src/waku.ts
@@ -3,7 +3,6 @@ import type { MultiaddrInput } from "@multiformats/multiaddr";
 
 import type { IConnectionManager } from "./connection_manager.js";
 import type { IFilter } from "./filter.js";
-import type { INextFilter } from "./filter_next.js";
 import type { IHealthIndicator } from "./health_indicator.js";
 import type { Libp2p } from "./libp2p.js";
 import type { ILightPush } from "./light_push.js";
@@ -36,12 +35,7 @@ export interface IWaku {
   relay?: IRelay;
   store?: IStore;
 
-  /**
-   * @deprecated use IWaku.nextFilter instead
-   */
   filter?: IFilter;
-
-  nextFilter?: INextFilter;
   lightPush?: ILightPush;
   connectionManager: IConnectionManager;
   health: IHealthIndicator;
@@ -217,7 +211,6 @@ export interface LightNode extends IWaku {
   relay: undefined;
   store: IStore;
   filter: IFilter;
-  nextFilter: INextFilter;
   lightPush: ILightPush;
 }
