--- packages/sdk/src/filter/constants.ts
@@ -1,3 +1,0 @@
-export const DEFAULT_KEEP_ALIVE = 60_000;
-export const DEFAULT_MAX_PINGS = 3;
-export const DEFAULT_LIGHT_PUSH_FILTER_CHECK = false;
