+++ packages/sdk/src/filter/filter.ts
@@ -1,77 +1,50 @@
 import { ConnectionManager, FilterCore } from "@waku/core";
 import type {
   Callback,
-  CreateSubscriptionResult,
   FilterProtocolOptions,
-  IAsyncIterator,
   IDecodedMessage,
   IDecoder,
   IFilter,
-  ILightPush,
-  Libp2p,
-  PubsubTopic,
-  SubscribeResult,
-  Unsubscribe
+  Libp2p
 } from "@waku/interfaces";
-import { NetworkConfig, ProtocolError } from "@waku/interfaces";
-import {
-  ensurePubsubTopicIsConfigured,
-  groupByContentTopic,
-  Logger,
-  shardInfoToPubsubTopics,
-  toAsyncIterator
-} from "@waku/utils";
+import { WakuMessage } from "@waku/proto";
+import { Logger } from "@waku/utils";
 
 import { PeerManager } from "../peer_manager/index.js";
 
 import { Subscription } from "./subscription.js";
-import { buildConfig } from "./utils.js";
+import { FilterConstructorParams } from "./types.js";
 
-const log = new Logger("sdk:filter");
+const log = new Logger("sdk:next-filter");
 
-type FilterConstructorParams = {
-  connectionManager: ConnectionManager;
-  libp2p: Libp2p;
-  peerManager: PeerManager;
-  lightPush?: ILightPush;
-  options?: Partial<FilterProtocolOptions>;
-};
+type PubsubTopic = string;
 
 export class Filter implements IFilter {
-  public readonly protocol: FilterCore;
+  private readonly libp2p: Libp2p;
+  private readonly protocol: FilterCore;
+  private readonly peerManager: PeerManager;
+  private readonly connectionManager: ConnectionManager;
 
   private readonly config: FilterProtocolOptions;
-  private connectionManager: ConnectionManager;
-  private libp2p: Libp2p;
-  private peerManager: PeerManager;
-  private lightPush?: ILightPush;
-  private activeSubscriptions = new Map<string, Subscription>();
+  private subscriptions = new Map<PubsubTopic, Subscription>();
 
   public constructor(params: FilterConstructorParams) {
-    this.config = buildConfig(params.options);
-    this.lightPush = params.lightPush;
-    this.peerManager = params.peerManager;
+    this.config = {
+      numPeersToUse: 2,
+      pingsBeforePeerRenewed: 3,
+      keepAliveIntervalMs: 60_000,
+      ...params.options
+    };
+
     this.libp2p = params.libp2p;
+    this.peerManager = params.peerManager;
     this.connectionManager = params.connectionManager;
 
     this.protocol = new FilterCore(
-      async (pubsubTopic, wakuMessage, peerIdStr) => {
-        const subscription = this.getActiveSubscription(pubsubTopic);
-        if (!subscription) {
-          log.error(
-            `No subscription locally registered for topic ${pubsubTopic}`
-          );
-          return;
-        }
-
-        await subscription.processIncomingMessage(wakuMessage, peerIdStr);
-      },
-
+      this.onIncomingMessage.bind(this),
       params.connectionManager.pubsubTopics,
       params.libp2p
     );
-
-    this.activeSubscriptions = new Map();
   }
 
   public get multicodec(): string {
@@ -79,219 +52,208 @@ export class Filter implements IFilter {
   }
 
   /**
-   * Opens a subscription with the Filter protocol using the provided decoders and callback.
-   * This method combines the functionality of creating a subscription and subscribing to it.
-   *
-   * @param {IDecoder<T> | IDecoder<T>[]} decoders - A single decoder or an array of decoders to use for decoding messages.
-   * @param {Callback<T>} callback - The callback function to be invoked with decoded messages.
-   *
-   * @returns {Promise<SubscribeResult>} A promise that resolves to an object containing:
-   *   - subscription: The created subscription object if successful, or null if failed.
-   *   - error: A ProtocolError if the subscription creation failed, or null if successful.
-   *   - results: An object containing arrays of failures and successes from the subscription process.
-   *     Only present if the subscription was created successfully.
-   *
-   * @throws {Error} If there's an unexpected error during the subscription process.
-   *
-   * @remarks
-   * This method attempts to create a subscription using the pubsub topic derived from the provided decoders,
-   * then tries to subscribe using the created subscription. The return value should be interpreted as follows:
-   * - If `subscription` is null and `error` is non-null, a critical error occurred and the subscription failed completely.
-   * - If `subscription` is non-null and `error` is null, the subscription was created successfully.
-   *   In this case, check the `results` field for detailed information about successes and failures during the subscription process.
-   * - Even if the subscription was created successfully, there might be some failures in the results.
+   * Unsubscribes from all active subscriptions across all pubsub topics.
    *
    * @example
-   * ```typescript
-   * const {subscription, error, results} = await waku.filter.subscribe(decoders, callback);
-   * if (!subscription || error) {
-   *   console.error("Failed to create subscription:", error);
-   * }
-   * console.log("Subscription created successfully");
-   * if (results.failures.length > 0) {
-   *   console.warn("Some errors occurred during subscription:", results.failures);
-   * }
-   * console.log("Successful subscriptions:", results.successes);
+   * // Clean up all subscriptions when React component unmounts
+   * useEffect(() => {
+   *   return () => filter.unsubscribeAll();
+   * }, [filter]);
    *
-   * ```
+   * @example
+   * // Reset subscriptions and start over
+   * filter.unsubscribeAll();
+   * await filter.subscribe(newDecoder, newCallback);
    */
-  public async subscribe<T extends IDecodedMessage>(
-    decoders: IDecoder<T> | IDecoder<T>[],
-    callback: Callback<T>
-  ): Promise<SubscribeResult> {
-    const uniquePubsubTopics = this.getUniquePubsubTopics(decoders);
-
-    if (uniquePubsubTopics.length !== 1) {
-      return {
-        subscription: null,
-        error: ProtocolError.INVALID_DECODER_TOPICS,
-        results: null
-      };
+  public unsubscribeAll(): void {
+    for (const subscription of this.subscriptions.values()) {
+      subscription.stop();
     }
 
-    const pubsubTopic = uniquePubsubTopics[0];
-
-    const { subscription, error } = await this.createSubscription(pubsubTopic);
-
-    if (error) {
-      return {
-        subscription: null,
-        error: error,
-        results: null
-      };
-    }
-
-    const { failures, successes } = await subscription.subscribe(
-      decoders,
-      callback
-    );
-    return {
-      subscription,
-      error: null,
-      results: {
-        failures: failures,
-        successes: successes
-      }
-    };
+    this.subscriptions.clear();
   }
 
   /**
-   * Creates a new subscription to the given pubsub topic.
-   * The subscription is made to multiple peers for decentralization.
-   * @param pubsubTopicShardInfo The pubsub topic to subscribe to.
-   * @returns The subscription object.
+   * Subscribes to messages with specified decoders and executes callback when a message is received.
+   * In case no peers available initially - will delay subscription till connects to any peer.
+   *
+   * @param decoders - Single decoder or array of decoders to subscribe to. All decoders must share the same pubsubTopic.
+   * @param callback - Function called when a message matching the decoder's contentTopic is received.
+   * @returns Promise that resolves to true if subscription was successful, false otherwise.
+   *
+   * @example
+   * // Subscribe to a single content topic
+   * await filter.subscribe(decoder, (msg) => console.log(msg));
+   *
+   * @example
+   * // Subscribe to multiple content topics with the same pubsub topic
+   * await filter.subscribe([decoder1, decoder2], (msg) => console.log(msg));
+   *
+   * @example
+   * // Handle subscription failure
+   * const success = await filter.subscribe(decoder, handleMessage);
+   * if (!success) {
+   *   console.error("Failed to subscribe");
+   * }
    */
-  private async createSubscription(
-    pubsubTopicShardInfo: NetworkConfig | PubsubTopic
-  ): Promise<CreateSubscriptionResult> {
-    const pubsubTopic =
-      typeof pubsubTopicShardInfo == "string"
-        ? pubsubTopicShardInfo
-        : shardInfoToPubsubTopics(pubsubTopicShardInfo)?.[0];
+  public async subscribe<T extends IDecodedMessage>(
+    decoder: IDecoder<T> | IDecoder<T>[],
+    callback: Callback<T>
+  ): Promise<boolean> {
+    const decoders = Array.isArray(decoder) ? decoder : [decoder];
 
-    ensurePubsubTopicIsConfigured(pubsubTopic, this.protocol.pubsubTopics);
+    if (decoders.length === 0) {
+      throw Error("Cannot subscribe with 0 decoders.");
+    }
 
-    const peerIds = this.peerManager.getPeers();
-    if (peerIds.length === 0) {
-      return {
-        error: ProtocolError.NO_PEER_AVAILABLE,
-        subscription: null
-      };
+    const pubsubTopics = decoders.map((v) => v.pubsubTopic);
+    const contentTopics = decoders.map((v) => v.contentTopic);
+
+    // doing this for simplicity, we can enable subscription for more than one PubsubTopic at once later when requested
+    if (!this.isSamePubsubTopic(decoders)) {
+      throw Error(
+        `Cannot subscribe to more than one pubsub topic at the same time, got pubsubTopics:${pubsubTopics}`
+      );
     }
 
     log.info(
-      `Creating filter subscription with ${peerIds.length} peers: `,
-      peerIds.map((id) => id.toString())
+      `Subscribing to content topic: ${contentTopics}, pubsub topic: ${pubsubTopics}`
     );
 
-    const subscription =
-      this.getActiveSubscription(pubsubTopic) ??
-      this.setActiveSubscription(
-        pubsubTopic,
-        new Subscription(
-          pubsubTopic,
-          this.protocol,
-          this.connectionManager,
-          this.peerManager,
-          this.libp2p,
-          this.config,
-          this.lightPush
-        )
-      );
+    const supportedPubsubTopic = this.connectionManager.pubsubTopics.includes(
+      pubsubTopics[0]
+    );
 
-    return {
-      error: null,
-      subscription
-    };
+    if (!supportedPubsubTopic) {
+      throw Error(
+        `Pubsub topic ${pubsubTopics[0]} has not been configured on this instance.`
+      );
+    }
+
+    let subscription = this.subscriptions.get(pubsubTopics[0]);
+    if (!subscription) {
+      subscription = new Subscription({
+        pubsubTopic: pubsubTopics[0],
+        libp2p: this.libp2p,
+        protocol: this.protocol,
+        config: this.config,
+        peerManager: this.peerManager
+      });
+      subscription.start();
+    }
+
+    const result = await subscription.add(decoders, callback);
+    this.subscriptions.set(pubsubTopics[0], subscription);
+
+    log.info(
+      `Subscription ${result ? "successful" : "failed"} for content topic: ${contentTopics}`
+    );
+
+    return result;
   }
 
   /**
-   * This method is used to satisfy the `IReceiver` interface.
+   * Unsubscribes from messages with specified decoders.
    *
-   * @hidden
+   * @param decoders - Single decoder or array of decoders to unsubscribe from. All decoders must share the same pubsubTopic.
+   * @returns Promise that resolves to true if unsubscription was successful, false otherwise.
    *
-   * @param decoders The decoders to use for the subscription.
-   * @param callback The callback function to use for the subscription.
-   * @param opts Optional protocol options for the subscription.
+   * @example
+   * // Unsubscribe from a single decoder
+   * await filter.unsubscribe(decoder);
    *
-   * @returns A Promise that resolves to a function that unsubscribes from the subscription.
+   * @example
+   * // Unsubscribe from multiple decoders at once
+   * await filter.unsubscribe([decoder1, decoder2]);
    *
-   * @remarks
-   * This method should not be used directly.
-   * Instead, use `createSubscription` to create a new subscription.
+   * @example
+   * // Handle unsubscription failure
+   * const success = await filter.unsubscribe(decoder);
+   * if (!success) {
+   *   console.error("Failed to unsubscribe");
+   * }
    */
-  public async subscribeWithUnsubscribe<T extends IDecodedMessage>(
-    decoders: IDecoder<T> | IDecoder<T>[],
-    callback: Callback<T>
-  ): Promise<Unsubscribe> {
-    const uniquePubsubTopics = this.getUniquePubsubTopics<T>(decoders);
-
-    if (uniquePubsubTopics.length === 0) {
-      throw Error(
-        "Failed to subscribe: no pubsubTopic found on decoders provided."
-      );
-    }
-
-    if (uniquePubsubTopics.length > 1) {
-      throw Error(
-        "Failed to subscribe: all decoders should have the same pubsub topic. Use createSubscription to be more agile."
-      );
-    }
-
-    const { subscription, error } = await this.createSubscription(
-      uniquePubsubTopics[0]
-    );
-
-    if (error) {
-      throw Error(`Failed to create subscription: ${error}`);
-    }
-
-    await subscription.subscribe(decoders, callback);
-
-    const contentTopics = Array.from(
-      groupByContentTopic(
-        Array.isArray(decoders) ? decoders : [decoders]
-      ).keys()
-    );
-
-    return async () => {
-      await subscription.unsubscribe(contentTopics);
-    };
-  }
-
-  public toSubscriptionIterator<T extends IDecodedMessage>(
-    decoders: IDecoder<T> | IDecoder<T>[]
-  ): Promise<IAsyncIterator<T>> {
-    return toAsyncIterator(this, decoders);
-  }
-
-  private getActiveSubscription(
-    pubsubTopic: PubsubTopic
-  ): Subscription | undefined {
-    return this.activeSubscriptions.get(pubsubTopic);
-  }
-
-  private setActiveSubscription(
-    pubsubTopic: PubsubTopic,
-    subscription: Subscription
-  ): Subscription {
-    this.activeSubscriptions.set(pubsubTopic, subscription);
-    return subscription;
-  }
-
-  private getUniquePubsubTopics<T extends IDecodedMessage>(
-    decoders: IDecoder<T> | IDecoder<T>[]
-  ): string[] {
-    if (!Array.isArray(decoders)) {
-      return [decoders.pubsubTopic];
-    }
+  public async unsubscribe<T extends IDecodedMessage>(
+    decoder: IDecoder<T> | IDecoder<T>[]
+  ): Promise<boolean> {
+    const decoders = Array.isArray(decoder) ? decoder : [decoder];
 
     if (decoders.length === 0) {
-      return [];
+      throw Error("Cannot unsubscribe with 0 decoders.");
     }
 
-    const pubsubTopics = new Set(decoders.map((d) => d.pubsubTopic));
+    const pubsubTopics = decoders.map((v) => v.pubsubTopic);
+    const contentTopics = decoders.map((v) => v.contentTopic);
 
-    return [...pubsubTopics];
+    // doing this for simplicity, we can enable unsubscribing with more than one PubsubTopic at once later when requested
+    if (!this.isSamePubsubTopic(decoders)) {
+      throw Error(
+        `Cannot unsubscribe with more than one pubsub topic at the same time, got pubsubTopics:${pubsubTopics}`
+      );
+    }
+
+    log.info(
+      `Unsubscribing from content topic: ${contentTopics}, pubsub topic: ${pubsubTopics}`
+    );
+
+    const supportedPubsubTopic = this.connectionManager.pubsubTopics.includes(
+      pubsubTopics[0]
+    );
+    if (!supportedPubsubTopic) {
+      throw Error(
+        `Pubsub topic ${pubsubTopics[0]} has not been configured on this instance.`
+      );
+    }
+
+    const subscription = this.subscriptions.get(pubsubTopics[0]);
+    if (!subscription) {
+      log.warn("No subscriptions associated with the decoder.");
+      return false;
+    }
+
+    const result = await subscription.remove(decoders);
+
+    if (subscription.isEmpty()) {
+      log.warn("Subscription has no decoders anymore, terminating it.");
+      subscription.stop();
+      this.subscriptions.delete(pubsubTopics[0]);
+    }
+
+    log.info(
+      `Unsubscribing ${result ? "successful" : "failed"} for content topic: ${contentTopics}`
+    );
+
+    return result;
+  }
+
+  private async onIncomingMessage(
+    pubsubTopic: string,
+    message: WakuMessage,
+    peerId: string
+  ): Promise<void> {
+    log.info(
+      `Received message for pubsubTopic:${pubsubTopic}, contentTopic:${message.contentTopic}, peerId:${peerId.toString()}`
+    );
+
+    const subscription = this.subscriptions.get(pubsubTopic);
+
+    if (!subscription) {
+      log.error(`No subscription locally registered for topic ${pubsubTopic}`);
+      return;
+    }
+
+    subscription.invoke(message, peerId);
+  }
+
+  private isSamePubsubTopic<T extends IDecodedMessage>(
+    decoders: IDecoder<T>[]
+  ): boolean {
+    const topics = new Set<string>();
+
+    for (const decoder of decoders) {
+      topics.add(decoder.pubsubTopic);
+    }
+
+    return topics.size === 1;
   }
 }
