+++ packages/sdk/src/filter/types.ts
@@ -1,12 +1,12 @@
 import { ConnectionManager } from "@waku/core";
 import { FilterCore } from "@waku/core";
-import type { Libp2p, NextFilterOptions } from "@waku/interfaces";
+import type { FilterProtocolOptions, Libp2p } from "@waku/interfaces";
 import { WakuMessage } from "@waku/proto";
 
 import { PeerManager } from "../peer_manager/index.js";
 
 export type FilterConstructorParams = {
-  options?: Partial<NextFilterOptions>;
+  options?: Partial<FilterProtocolOptions>;
   libp2p: Libp2p;
   peerManager: PeerManager;
   connectionManager: ConnectionManager;
@@ -20,6 +20,6 @@ export type SubscriptionParams = {
   libp2p: Libp2p;
   pubsubTopic: string;
   protocol: FilterCore;
-  config: NextFilterOptions;
+  config: FilterProtocolOptions;
   peerManager: PeerManager;
 };
