--- packages/sdk/src/filter_next/index.ts
@@ -1,1 +1,0 @@
-export { Filter as NextFilter } from "./filter.js";
