+++ packages/tests/tests/filter/subscribe.node.spec.ts
@@ -1,5 +1,5 @@
-import { createDecoder, createEncoder, DecodedMessage } from "@waku/core";
-import { IDecoder, LightNode } from "@waku/interfaces";
+import { createDecoder, createEncoder } from "@waku/core";
+import { IDecodedMessage, IDecoder, LightNode } from "@waku/interfaces";
 import {
   ecies,
   generatePrivateKey,
@@ -39,7 +39,7 @@ import {
 } from "./utils.js";
 
 const runTests = (strictCheckNodes: boolean): void => {
-  describe(`Waku Filter V2: Subscribe: Multiple Service Nodes: Strict Check mode: ${strictCheckNodes}`, function () {
+  describe(`Waku Filter Next: Subscribe: Multiple Service Nodes: Strict Check mode: ${strictCheckNodes}`, function () {
     this.timeout(100000);
     let waku: LightNode;
     let serviceNodes: ServiceNodesFleet;
@@ -60,8 +60,8 @@ const runTests = (strictCheckNodes: boolean): void => {
     it("Subscribe and receive messages via lightPush", async function () {
       expect(waku.libp2p.getConnections()).has.length(2);
 
-      await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
 
@@ -92,8 +92,8 @@ const runTests = (strictCheckNodes: boolean): void => {
         TestPubsubTopic
       );
 
-      await waku.filter.subscribe(
-        [decoder],
+      await waku.nextFilter.subscribe(
+        decoder,
         serviceNodes.messageCollector.callback
       );
 
@@ -125,8 +125,8 @@ const runTests = (strictCheckNodes: boolean): void => {
         TestPubsubTopic
       );
 
-      await waku.filter.subscribe(
-        [decoder],
+      await waku.nextFilter.subscribe(
+        decoder,
         serviceNodes.messageCollector.callback
       );
 
@@ -146,8 +146,8 @@ const runTests = (strictCheckNodes: boolean): void => {
     });
 
     it("Subscribe and receive messages via waku relay post", async function () {
-      await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
 
@@ -173,8 +173,8 @@ const runTests = (strictCheckNodes: boolean): void => {
     });
 
     it("Subscribe and receive 2 messages on the same topic", async function () {
-      await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
 
@@ -208,8 +208,8 @@ const runTests = (strictCheckNodes: boolean): void => {
 
     it("Subscribe and receive messages on 2 different content topics", async function () {
       // Subscribe to the first content topic and send a message.
-      await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
       await waku.lightPush.send(TestEncoder, messagePayload);
@@ -231,8 +231,8 @@ const runTests = (strictCheckNodes: boolean): void => {
         pubsubTopic: TestPubsubTopic
       });
       const newDecoder = createDecoder(newContentTopic, TestPubsubTopic);
-      await waku.filter.subscribe(
-        [newDecoder],
+      await waku.nextFilter.subscribe(
+        newDecoder,
         serviceNodes.messageCollector.callback
       );
       await waku.lightPush.send(newEncoder, {
@@ -267,8 +267,8 @@ const runTests = (strictCheckNodes: boolean): void => {
 
       // Subscribe to all 20 topics.
       for (let i = 0; i < topicCount; i++) {
-        await waku.filter.subscribe(
-          [td.decoders[i]],
+        await waku.nextFilter.subscribe(
+          td.decoders[i],
           serviceNodes.messageCollector.callback
         );
       }
@@ -293,12 +293,45 @@ const runTests = (strictCheckNodes: boolean): void => {
       });
     });
 
+    // skip for now, will be enabled once old Filter is removed as it exausts amount of streams avaialble
+    it.skip("Subscribe to 30 topics in separate streams (30 streams for Filter is limit) at once and receives messages", async function () {
+      this.timeout(100_000);
+      const topicCount = 30;
+      const td = generateTestData(topicCount, { pubsubTopic: TestPubsubTopic });
+
+      for (let i = 0; i < topicCount; i++) {
+        await waku.nextFilter.subscribe(
+          td.decoders[i],
+          serviceNodes.messageCollector.callback
+        );
+      }
+
+      // Send a unique message on each topic.
+      for (let i = 0; i < topicCount; i++) {
+        await waku.lightPush.send(td.encoders[i], {
+          payload: utf8ToBytes(`Message for Topic ${i + 1}`)
+        });
+      }
+
+      // Verify that each message was received on the corresponding topic.
+      expect(
+        await serviceNodes.messageCollector.waitForMessages(topicCount)
+      ).to.eq(true);
+      td.contentTopics.forEach((topic, index) => {
+        serviceNodes.messageCollector.verifyReceivedMessage(index, {
+          expectedContentTopic: topic,
+          expectedMessageText: `Message for Topic ${index + 1}`,
+          expectedPubsubTopic: TestPubsubTopic
+        });
+      });
+    });
+
     it("Subscribe to 100 topics (new limit) at once and receives messages", async function () {
       this.timeout(100_000);
       const topicCount = 100;
       const td = generateTestData(topicCount, { pubsubTopic: TestPubsubTopic });
 
-      await waku.filter.subscribe(
+      await waku.nextFilter.subscribe(
         td.decoders,
         serviceNodes.messageCollector.callback
       );
@@ -328,19 +361,10 @@ const runTests = (strictCheckNodes: boolean): void => {
       const td = generateTestData(topicCount, { pubsubTopic: TestPubsubTopic });
 
       try {
-        const { error, results } = await waku.filter.subscribe(
+        await waku.nextFilter.subscribe(
           td.decoders,
           serviceNodes.messageCollector.callback
         );
-        if (error) {
-          throw error;
-        }
-        const { failures, successes } = results;
-        if (failures.length === 0 || successes.length > 0) {
-          throw new Error(
-            `Subscribe to ${topicCount} topics was successful but was expected to fail with a specific error.`
-          );
-        }
       } catch (err) {
         if (
           err instanceof Error &&
@@ -361,19 +385,19 @@ const runTests = (strictCheckNodes: boolean): void => {
       const td1 = generateTestData(topicCount1, {
         pubsubTopic: TestPubsubTopic
       });
+
       const topicCount2 = 4;
       const td2 = generateTestData(topicCount2, {
         pubsubTopic: TestPubsubTopic
       });
 
-      // Subscribe to the first set of topics.
-      await waku.filter.subscribe(
+      await waku.nextFilter.subscribe(
         td1.decoders,
         serviceNodes.messageCollector.callback
       );
 
       // Subscribe to the second set of topics which has overlapping topics with the first set.
-      await waku.filter.subscribe(
+      await waku.nextFilter.subscribe(
         td2.decoders,
         serviceNodes.messageCollector.callback
       );
@@ -394,23 +418,22 @@ const runTests = (strictCheckNodes: boolean): void => {
         });
       }
 
-      // Check if all messages were received.
-      // Since there are overlapping topics, there should be 6 messages in total (2 from the first set + 4 from the second set).
+      // Since there are overlapping topics, there should be 10 messages in total because overlaping decoders handle them
       expect(
-        await serviceNodes.messageCollector.waitForMessages(6, { exact: true })
+        await serviceNodes.messageCollector.waitForMessages(10, { exact: true })
       ).to.eq(true);
     });
 
     it("Refresh subscription", async function () {
-      await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
       await waku.lightPush.send(TestEncoder, { payload: utf8ToBytes("M1") });
 
       // Resubscribe (refresh) to the same topic and send another message.
-      await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
       await waku.lightPush.send(TestEncoder, { payload: utf8ToBytes("M2") });
@@ -449,8 +472,8 @@ const runTests = (strictCheckNodes: boolean): void => {
           }
         });
 
-        await waku.filter.subscribe(
-          [newDecoder as IDecoder<DecodedMessage>],
+        await waku.nextFilter.subscribe(
+          newDecoder as IDecoder<IDecodedMessage>,
           serviceNodes.messageCollector.callback
         );
         await waku.lightPush.send(newEncoder, messagePayload);
@@ -467,8 +490,8 @@ const runTests = (strictCheckNodes: boolean): void => {
     });
 
     it("Add multiple subscription objects on single nwaku node", async function () {
-      await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
       await waku.lightPush.send(TestEncoder, { payload: utf8ToBytes("M1") });
@@ -479,8 +502,8 @@ const runTests = (strictCheckNodes: boolean): void => {
         pubsubTopic: TestPubsubTopic
       });
       const newDecoder = createDecoder(newContentTopic, TestPubsubTopic);
-      await waku.filter.subscribe(
-        [newDecoder],
+      await waku.nextFilter.subscribe(
+        newDecoder,
         serviceNodes.messageCollector.callback
       );
 
@@ -506,8 +529,8 @@ const runTests = (strictCheckNodes: boolean): void => {
       // setup check
       expect(waku.libp2p.getConnections()).has.length(2);
 
-      await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
 
@@ -549,8 +572,8 @@ const runTests = (strictCheckNodes: boolean): void => {
     });
 
     it("Subscribe and receive messages from 2 nwaku nodes each with different pubsubtopics", async function () {
-      await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
 
@@ -582,8 +605,8 @@ const runTests = (strictCheckNodes: boolean): void => {
 
         const messageCollector2 = new MessageCollector();
 
-        await waku.filter.subscribe(
-          [customDecoder],
+        await waku.nextFilter.subscribe(
+          customDecoder,
           messageCollector2.callback
         );
 
@@ -629,8 +652,8 @@ const runTests = (strictCheckNodes: boolean): void => {
 
       // this subscription object is set up with the `customPubsubTopic1` but we're passing it a Decoder with the `customPubsubTopic2`
       try {
-        await waku.filter.subscribe(
-          [wrongDecoder],
+        await waku.nextFilter.subscribe(
+          wrongDecoder,
           serviceNodes.messageCollector.callback
         );
       } catch (error) {
