+++ packages/tests/tests/filter/unsubscribe.node.spec.ts
@@ -23,7 +23,7 @@ import {
 } from "./utils.js";
 
 const runTests = (strictCheckNodes: boolean): void => {
-  describe(`Waku Filter V2: Unsubscribe: Multiple Nodes: Strict Checking: ${strictCheckNodes}`, function () {
+  describe(`Waku Filter Next: Unsubscribe: Multiple Nodes: Strict Checking: ${strictCheckNodes}`, function () {
     // Set the timeout for all tests in this suite. Can be overwritten at test level
     this.timeout(10000);
     let waku: LightNode;
@@ -45,26 +45,22 @@ const runTests = (strictCheckNodes: boolean): void => {
     });
 
     it("Unsubscribe 1 topic - node subscribed to 1 topic", async function () {
-      const { error, subscription } = await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
-      if (error) {
-        throw error;
-      }
+
       await waku.lightPush.send(TestEncoder, messagePayload);
       expect(await serviceNodes.messageCollector.waitForMessages(1)).to.eq(
         true
       );
 
-      // Unsubscribe from the topic and send again
-      await subscription.unsubscribe([TestContentTopic]);
+      await waku.nextFilter.unsubscribe(TestDecoder);
       await waku.lightPush.send(TestEncoder, { payload: utf8ToBytes("M1") });
       expect(await serviceNodes.messageCollector.waitForMessages(2)).to.eq(
         false
       );
 
-      // Check that from 2 messages send only the 1st was received
       serviceNodes.messageCollector.verifyReceivedMessage(0, {
         expectedMessageText: messageText,
         expectedContentTopic: TestContentTopic
@@ -76,21 +72,19 @@ const runTests = (strictCheckNodes: boolean): void => {
 
     it("Unsubscribe 1 topic - node subscribed to 2 topics", async function () {
       // Subscribe to 2 topics and send messages
-      const { error, subscription } = await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
-      if (error) {
-        throw error;
-      }
+
       const newContentTopic = "/test/2/waku-filter";
       const newEncoder = createEncoder({
         contentTopic: newContentTopic,
         pubsubTopic: TestPubsubTopic
       });
       const newDecoder = createDecoder(newContentTopic, TestPubsubTopic);
-      await waku.filter.subscribe(
-        [newDecoder],
+      await waku.nextFilter.subscribe(
+        newDecoder,
         serviceNodes.messageCollector.callback
       );
       await waku.lightPush.send(TestEncoder, { payload: utf8ToBytes("M1") });
@@ -100,7 +94,7 @@ const runTests = (strictCheckNodes: boolean): void => {
       );
 
       // Unsubscribe from the first topic and send again
-      await subscription.unsubscribe([TestContentTopic]);
+      await waku.nextFilter.unsubscribe(TestDecoder);
       await waku.lightPush.send(TestEncoder, { payload: utf8ToBytes("M3") });
       await waku.lightPush.send(newEncoder, { payload: utf8ToBytes("M4") });
       expect(await serviceNodes.messageCollector.waitForMessages(3)).to.eq(
@@ -114,8 +108,8 @@ const runTests = (strictCheckNodes: boolean): void => {
 
     it("Unsubscribe 2 topics - node subscribed to 2 topics", async function () {
       // Subscribe to 2 topics and send messages
-      await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
       const newContentTopic = "/test/2/waku-filter";
@@ -124,13 +118,11 @@ const runTests = (strictCheckNodes: boolean): void => {
         pubsubTopic: TestPubsubTopic
       });
       const newDecoder = createDecoder(newContentTopic, TestPubsubTopic);
-      const { error, subscription } = await waku.filter.subscribe(
-        [newDecoder],
+      await waku.nextFilter.subscribe(
+        newDecoder,
         serviceNodes.messageCollector.callback
       );
-      if (error) {
-        throw error;
-      }
+
       await waku.lightPush.send(TestEncoder, { payload: utf8ToBytes("M1") });
       await waku.lightPush.send(newEncoder, { payload: utf8ToBytes("M2") });
       expect(await serviceNodes.messageCollector.waitForMessages(2)).to.eq(
@@ -138,7 +130,8 @@ const runTests = (strictCheckNodes: boolean): void => {
       );
 
       // Unsubscribe from both and send again
-      await subscription.unsubscribe([TestContentTopic, newContentTopic]);
+      await waku.nextFilter.unsubscribe(TestDecoder);
+      await waku.nextFilter.unsubscribe(newDecoder);
       await waku.lightPush.send(TestEncoder, { payload: utf8ToBytes("M3") });
       await waku.lightPush.send(newEncoder, { payload: utf8ToBytes("M4") });
       expect(await serviceNodes.messageCollector.waitForMessages(3)).to.eq(
@@ -152,13 +145,11 @@ const runTests = (strictCheckNodes: boolean): void => {
 
     it("Unsubscribe topics the node is not subscribed to", async function () {
       // Subscribe to 1 topic and send message
-      const { error, subscription } = await waku.filter.subscribe(
-        [TestDecoder],
+      await waku.nextFilter.subscribe(
+        TestDecoder,
         serviceNodes.messageCollector.callback
       );
-      if (error) {
-        throw error;
-      }
+
       await waku.lightPush.send(TestEncoder, { payload: utf8ToBytes("M1") });
       expect(await serviceNodes.messageCollector.waitForMessages(1)).to.eq(
         true
@@ -167,8 +158,9 @@ const runTests = (strictCheckNodes: boolean): void => {
       expect(serviceNodes.messageCollector.count).to.eq(1);
 
       // Unsubscribe from topics that the node is not not subscribed to and send again
-      await subscription.unsubscribe([]);
-      await subscription.unsubscribe(["/test/2/waku-filter"]);
+      await waku.nextFilter.unsubscribe(
+        createDecoder("/test/2/waku-filter", TestDecoder.pubsubTopic)
+      );
       await waku.lightPush.send(TestEncoder, { payload: utf8ToBytes("M2") });
       expect(await serviceNodes.messageCollector.waitForMessages(2)).to.eq(
         true
@@ -179,66 +171,42 @@ const runTests = (strictCheckNodes: boolean): void => {
       await serviceNodes.confirmMessageLength(2);
     });
 
-    it("Unsubscribes all - node subscribed to 1 topic", async function () {
-      const { error, subscription } = await waku.filter.subscribe(
-        [TestDecoder],
-        serviceNodes.messageCollector.callback
-      );
-      if (error) {
-        throw error;
-      }
-      await waku.lightPush.send(TestEncoder, { payload: utf8ToBytes("M1") });
-      expect(await serviceNodes.messageCollector.waitForMessages(1)).to.eq(
-        true
-      );
-      expect(serviceNodes.messageCollector.count).to.eq(1);
-
-      // Unsubscribe from all topics and send again
-      await subscription.unsubscribeAll();
-      await waku.lightPush.send(TestEncoder, { payload: utf8ToBytes("M2") });
-      expect(await serviceNodes.messageCollector.waitForMessages(2)).to.eq(
-        false
-      );
-
-      // Check that from 2 messages send only the 1st was received
-      expect(serviceNodes.messageCollector.count).to.eq(1);
-      await serviceNodes.confirmMessageLength(2);
-    });
-
-    it("Unsubscribes all - node subscribed to 10 topics", async function () {
-      // Subscribe to 10 topics and send message
-      const topicCount = 10;
+    it("Unsubscribe from 100 topics (new limit) at once and receives messages", async function () {
+      this.timeout(100_000);
+      const topicCount = 100;
       const td = generateTestData(topicCount, { pubsubTopic: TestPubsubTopic });
-      const { error, subscription } = await waku.filter.subscribe(
+
+      await waku.nextFilter.subscribe(
         td.decoders,
         serviceNodes.messageCollector.callback
       );
-      if (error) {
-        throw error;
-      }
+
       for (let i = 0; i < topicCount; i++) {
         await waku.lightPush.send(td.encoders[i], {
-          payload: utf8ToBytes(`M${i + 1}`)
+          payload: utf8ToBytes(`Message for Topic ${i + 1}`)
         });
       }
-      expect(await serviceNodes.messageCollector.waitForMessages(10)).to.eq(
-        true
-      );
 
-      // Unsubscribe from all topics and send again
-      await subscription.unsubscribeAll();
+      expect(
+        await serviceNodes.messageCollector.waitForMessages(topicCount)
+      ).to.eq(true);
+      td.contentTopics.forEach((topic, index) => {
+        serviceNodes.messageCollector.verifyReceivedMessage(index, {
+          expectedContentTopic: topic,
+          expectedMessageText: `Message for Topic ${index + 1}`,
+          expectedPubsubTopic: TestPubsubTopic
+        });
+      });
+
+      await waku.nextFilter.unsubscribe(td.decoders);
+
       for (let i = 0; i < topicCount; i++) {
         await waku.lightPush.send(td.encoders[i], {
-          payload: utf8ToBytes(`M${topicCount + i + 1}`)
+          payload: utf8ToBytes(`Message for Topic ${i + 1}`)
         });
       }
-      expect(await serviceNodes.messageCollector.waitForMessages(11)).to.eq(
-        false
-      );
 
-      // Check that from 20 messages send only 10 were received
-      expect(serviceNodes.messageCollector.count).to.eq(10);
-      await serviceNodes.confirmMessageLength(20);
+      expect(serviceNodes.messageCollector.count).to.eq(100);
     });
   });
 };
